# Make otimesu and otimesl faster for symmetric tensors