--- a/src/tensor_products.jl
+++ b/src/tensor_products.jl
@@ -184,8 +184,8 @@ julia> otimesu(A, B)
 ```
 """
 @inline function otimesu(S1::SecondOrderTensor{dim}, S2::SecondOrderTensor{dim}) where {dim}
-    S1_ = Tensor{2,dim}(S1) # Convert to full tensor if symmetric to make 10x faster...
-    S2_ = Tensor{2,dim}(S2)
+    S1_ = convert(Tensor{2,dim}, S1) # Convert to full tensor if symmetric to make 10x faster... (see Tensors.jl#164)
+    S2_ = convert(Tensor{2,dim}, S2)
     return Tensor{4, dim}((i,j,k,l) -> S1_[i,k] * S2_[j,l])
 end