--- a/src/tensor_products.jl
+++ b/src/tensor_products.jl
@@ -184,8 +184,8 @@ julia> otimesu(A, B)
 ```
 """
 @inline function otimesu(S1::SecondOrderTensor{dim}, S2::SecondOrderTensor{dim}) where {dim}
-    S1_ = convert(Tensor{2,dim}, S1) # Convert to full tensor if symmetric to make 10x faster... (see Tensors.jl#164)
-    S2_ = convert(Tensor{2,dim}, S2)
+    S1_ = convert(Tensor, S1) # Convert to full tensor if symmetric to make 10x faster... (see Tensors.jl#164)
+    S2_ = convert(Tensor, S2)
     return Tensor{4, dim}((i,j,k,l) -> S1_[i,k] * S2_[j,l])
 end
 
@@ -220,8 +220,8 @@ julia> otimesl(A, B)
 ```
 """
 @inline function otimesl(S1::SecondOrderTensor{dim}, S2::SecondOrderTensor{dim}) where {dim}
-    S1_ = convert(Tensor{2,dim}, S1) # Convert to full tensor if symmetric to make 10x faster... (see Tensors.jl#164)
-    S2_ = convert(Tensor{2,dim}, S2)
+    S1_ = convert(Tensor, S1) # Convert to full tensor if symmetric to make 10x faster... (see Tensors.jl#164)
+    S2_ = convert(Tensor, S2)
     return Tensor{4, dim}((i,j,k,l) -> S1_[i,l] * S2_[j,k])
 end