See #164 

## Timings 
Too large variance to determine if this has any negative effect when giving full tensors), but doesn't seem to have any significant effect... 

```julia
a, b = (rand(Tensor{2,3}) for _ in 1:2);
as, bs = (rand(SymmetricTensor{2,3}) for _ in 1:2);`
```

Command | Master | PR
--- | --- | ---
`@btime otimesu($a, $b);` | 13.627 ns | 10.611 ns | 
`@btime otimesu($as, $bs);` | 334.421 ns |  20.641 ns | 
`@btime otimesl($a, $b);` | 18.437 ns | 11.700 ns | 
`@btime otimesl($as, $bs);` | 343.071 ns | 25.451 ns |