@KnutAM 2021-11-03
Nice, that should be guaranteed to remove any overhead!