@fredrikekre 2021-11-03
No SIMD and lots of data shuffling in the slow case.